+++ scripts/cljs.sh
@@ -1,3 +1,4 @@
 #! /bin/bash
 
-clj -R:figwheel -m figwheel.main --build twitch --repl
+# How to run multiple builds?
+clj -R:figwheel -m figwheel.main --build emoticons --repl