+++ src/clj/upgrade/http.clj
@@ -1,14 +1,16 @@
 (ns upgrade.http
-  (:require [clojure.java.io :as io]
-            [upgrade.common :refer [log]]
-            [org.httpkit.server :as httpkit]
-            [bidi.bidi :refer [match-route path-for]]
+  (:require [bidi.bidi :refer [match-route path-for]]
             [bidi.ring :refer (make-handler)]
+            [buddy.sign.jwt :as jwt]
+            [clojure.java.io :as io]
+            [clj-time.core :as time]
+            [org.httpkit.server :as httpkit :refer [send! with-channel]]
             [ring.util.response :as res]
             [ring.middleware.json :refer [wrap-json-response]]
-            [ring.mock.request :as mock]))
+            [ring.mock.request :as mock]
+            [upgrade.common :refer [log]]))
 
-(defonce http-state (atom {:color "purple"}))
+(defonce http-state (atom "purple"))
 
 (def available-colors {0 "purple"
                        1 "red"
@@ -28,15 +30,22 @@
       (get available-colors (inc idx)))))
 
 (defn color-query-handler
-  [request]
   "Return the current color. wrap-json-response will automatically
   convert clojure in body into json"
+  [request]
+  (log "HTTPKIT: Attempting to get current color")
   (res/response @http-state))
 
 (defn next-color-handler
-  [request]
   "Return the next color."
-  (swap! http-state assoc :color (next-color (:color @http-state) available-colors))
+  [request]
+  (log "HTTPKIT: Attempting to find next color")
+  (log request)
+  (let [headers (:headers request)
+        auth-header (get headers "authorization")]
+    (log auth-header)
+    )
+  (reset! http-state (next-color @http-state available-colors))
   (res/response @http-state))
 
 (defn resources-handler
@@ -49,16 +58,42 @@
       (res/response "file not found"))
     ))
 
+(defonce ws-clients (atom #{}))
+
+(defn websocket-handler [request]
+  (with-channel request channel
+    (swap! ws-clients conj channel)))
+
 (def routes
   ["/" {"color/query" color-query-handler
         "color/cycle" next-color-handler
         #".+\.html" resources-handler
         #"js/.+" resources-handler
+        "ws" websocket-handler 
         }])
+
+;; Access-Control-Allow-Origin
+(defn wrap-extra-headers [handler extra-headers]
+  (fn [request]
+    (let [response (handler request)]
+         (update-in response [:headers]
+                    (fn [headers] (merge headers extra-headers))))))
 
 (def handler
   (wrap-json-response
-   (make-handler routes)))
+   (wrap-extra-headers
+    (make-handler routes)
+    ;; remove all this for production,
+    ;; but we need it for the developer rig
+    {"Access-Control-Allow-Origin" "*"
+     "Access-Control-Allow-Credentials", "true"
+     "Access-Control-Allow-Methods", "GET,HEAD,OPTIONS,POST,PUT"
+     "Access-Control-Allow-Headers", (str "Access-Control-Allow-Headers, "
+                                          "Origin,Accept, X-Requested-With, "
+                                          "Content-Type, "
+                                          "Access-Control-Request-Method, "
+                                          "Access-Control-Request-Headers, "
+                                          "Authorization")})))
 
 (defn app [req]
   (handler req))